USRP Header library
- Safe USRP handling with errors and graceful exits
- Create single USRP object and modify parameters
- Getters and setters to modify USRP parameters (with range checks)
- Test file to verify USRP header library

Point of this library is to learn how we can use the UHD library. You'll
notice that the setters are basically single function calls, i.e.
usrp_device_->set_rx_gain(), with fluff around it to validate the device
and gracefully reset if needed. This hopefully would prevent bricking
and malfunctioning of the USRP board, super useful in my case since im
remoting into the machine and prefer not to drive to the lab everytime i
need to reset the USRPs :]

Left is multithreading to support receiving samples and drawing on ImGui
and of course the deep learning model (including training, testing).
Will use the same method in my research project (Journal & thesis) to
train the model.